add mongodb

+++ src/controllers/paleta.controller.js
@@ -1,33 +1,79 @@
 const paletasService = require('../services/paleta.service');
+const mongoose = require('mongoose');
 
-const findAllPaletasController = (req, res) => {
-  const paletas = paletasService.findAllPaletasService();
-
-  res.send(paletas);
+const findAllPaletasController = async (req, res) => {
+  const allPaletas = await paletasService.findAllPaletasService();
+  if (allPaletas.length == 0) {
+    return res
+      .status(404)
+      .send({ message: 'Não existe nenhuma paleta cadastrada!' });
+  }
+  res.send(allPaletas);
 };
 
-const findPaletaByIdController = (req, res) => {
-  const idParam = Number(req.params.id);
-  const chosenPaleta = paletasService.findPaletaByIdService(idParam);
+const findPaletaByIdController = async (req, res) => {
+  const idParam = (req.params.id);
+
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    return res.status(400).send({ message: 'Id inválido!'})
+  }
+
+  const chosenPaleta = await paletasService.findPaletaByIdService(idParam);
+
+  if (!chosenPaleta) {
+    return res.status(404).send({ message: 'Paleta não encontrada!' });
+  }
+
   res.send(chosenPaleta);
 };
 
-const createPaletaController = (req, res) => {
+const createPaletaController = async (req, res) => {
   const paleta = req.body;
-  const newPaleta = paletasService.createPaletaService(paleta);
+
+  if (
+    !paleta ||
+    !paleta.sabor ||
+    !paleta.descricao ||
+    !paleta.foto ||
+    !paleta.preco
+  ) {
+    return res.status(400).send({
+      message:
+        'Você não preencheu todos os dados para adicionar uma nova paleta ao cardápio!',
+    });
+  }
+
+  const newPaleta = await paletasService.createPaletaService(paleta);
+    
   res.send(newPaleta);
 };
 
-const updatePaletaController = (req, res) => {
-  const idParam = Number(req.params.id);
+const updatePaletaController = async (req, res) => {
+  const idParam = (req.params.id);
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    return res.status(400).send({ message: 'Id inválido!'})
+  }
   const paletaEdit = req.body;
-  const updatePaleta = paletasService.updatePaletaService(idParam, paletaEdit);
+  const updatePaleta = await paletasService.updatePaletaService(idParam, paletaEdit);
   res.send(updatePaleta);
 };
 
-const deletePaletaController = (req, res) => {
+const deletePaletaController = async (req, res) => {
   const idParam = req.params.id;
-  paletasService.deletePaletaService(idParam);
+
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    res.status(400).send({ message: 'ID inválido!' });
+    return;
+  }
+
+  const chosenPaleta = await paletasService.findPaletaByIdService(idParam);
+
+  if (!chosenPaleta) {
+    return res.status(404).send({ message: 'Paleta não encontrada!' });
+  }
+
+  await paletasService.deletePaletaService(idParam);
+
   res.send({ message: 'Paleta deletada com sucesso!' });
 };
